requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
redis==5.0.1
celery==5.3.4
boto3==1.34.0
paramiko==3.4.0
ldap3==2.9.1
pymongo==4.6.0
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
elasticsearch==8.11.0
pandas==2.1.4
numpy==1.25.2
Pillow==10.1.0
beautifulsoup4==4.12.2 